:hammer: Refactor tests
